+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Improve caching of parse and validation results for performance
 - Update formatting
 - Fix compatibility issue with `prettier@3.x` (`format` became async)
+- Add [audit](https://github.com/graphql/graphql-http) for [GraphQL over HTTP spec](https://graphql.github.io/graphql-over-http/)
 - Update deps
 
 ## 1.8.1